[ALPS03793590] VCM: Unintentional integer overflow
[Detail] Fix Unintentional integer overflow

[Solution]

MTK-Commit-Id: 629a3f26ed55dccb433f87dcdfdcd4089637c415

CR-Id: ALPS03793590
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/lens/main/common/lc898217af/LC898217AF.c
@@ -350,8 +350,10 @@ int LC898217AF_Release(struct inode *a_pstInode, struct file *a_pstFile)
 		unsigned int diff_ms;
 
 		g_TSAFClose = CURRENT_TIME;
-		start_ms = g_TSAFOpen.tv_sec * 1000 + (g_TSAFOpen.tv_nsec / 1000000);
-		end_ms = g_TSAFClose.tv_sec * 1000 + (g_TSAFClose.tv_nsec / 1000000);
+		start_ms = (unsigned long long)g_TSAFOpen.tv_sec * 1000L +
+						(g_TSAFOpen.tv_nsec / 1000000);
+		end_ms = (unsigned long long)g_TSAFClose.tv_sec * 1000L +
+						(g_TSAFClose.tv_nsec / 1000000);
 		diff_ms = end_ms - start_ms;
 
 		LOG_INF("Wait - Excute Time %d , FirstAFUninit(%d)\n", diff_ms, g_FirstAFUninit);
